% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_historical_excel.R
\name{read_historical_excel}
\alias{read_historical_excel}
\title{Parse the spreadsheet that contains the census's history of poverty levels}
\usage{
read_historical_excel(.path)
}
\arguments{
\item{.path}{\if{html}{\out{<chr>}} the full path to the file}
}
\value{
<data.frame> a table of poverty levels
}
\description{
Parse the spreadsheet that contains the census's history of poverty levels
}
